fix(frontend/copilot): address review on the token devtool
Correctness:
- ChatContainer's breakdown guard tracked only a message count, so
  switching to a thread with the same number of messages skipped the
  recompute and left the badge with a stale/empty breakdown. Key the
  guard by session as well.
- trimZeros stripped real digits from integer strings ("100" -> "1").
  Only trim when there is a fractional part. Latent today since every
  caller passes toFixed() output, but the name invites reuse.
- TurnRow keys added a timestamp to an index, so at=1000,i=1 and
  at=1001,i=0 collided. Use a composite key.

Stability:
- The tee'd tap kept reading after the consumer aborted, holding the
  connection open past cancellation (a tee source is only released once
  both branches let go). Cancel the tap on init.signal abort. Cancelled
  through the reader, since scanForUsage holds the lock and calling
  tap.cancel() on a locked stream throws.

Structure: both new files were over the ~200-line guideline, so
tokenDevtool.ts is split into gate/tokenMath/store/usageTap and the
badge's sub-components move into its local components/ folder with
pct + BREAKDOWN_COLORS in helpers.ts. No barrel files.

Adds a regression test for the abort path (verified failing without the
fix) and pins formatTokenCount(100_000) === "100k".
