+++ autogpt_platform/frontend/src/app/(platform)/copilot/components/TokenDevtoolBadge/TokenDevtoolBadge.tsx
@@ -8,18 +8,18 @@ import {
 } from "@/components/molecules/Popover/Popover";
 import { DashboardSpeed02Icon } from "@hugeicons/core-free-icons";
 import { useState } from "react";
+import { useTokenDevtoolStore } from "../../tokenDevtool/store";
 import {
   AUTOCOMPACT_TOKENS,
-  BASE_CONTEXT_ESTIMATE,
   MODEL_CONTEXT_WINDOW,
   breakdownTotal,
   displayContext,
   formatTokenCount,
-  turnInputTokens,
-  useTokenDevtoolStore,
-  type ContextBreakdown,
-  type TokenTurn,
-} from "../../tokenDevtool";
+} from "../../tokenDevtool/tokenMath";
+import { BreakdownSection } from "./components/BreakdownSection";
+import { ContextBar } from "./components/ContextBar";
+import { MiniBar } from "./components/MiniBar";
+import { TurnRow } from "./components/TurnRow";
 
 interface Props {
   sessionId: string;
@@ -85,139 +85,11 @@ export function TokenDevtoolBadge({ sessionId }: Props) {
         ) : (
           <div className="flex max-h-40 flex-col gap-1 overflow-y-auto">
             {turns.map((turn, i) => (
-              <TurnRow key={turn.at + i} index={i} turn={turn} />
+              <TurnRow key={`${turn.at}-${i}`} index={i} turn={turn} />
             ))}
           </div>
         )}
       </PopoverContent>
     </Popover>
-  );
-}
-
-function pct(context: number): number {
-  return Math.min(100, (context / MODEL_CONTEXT_WINDOW) * 100);
-}
-
-function MiniBar({ context }: { context: number }) {
-  return (
-    <span className="relative h-1 w-10 overflow-hidden rounded-full bg-zinc-200">
-      <span
-        className={
-          "absolute inset-y-0 left-0 rounded-full " +
-          (context >= AUTOCOMPACT_TOKENS ? "bg-amber-400" : "bg-zinc-500")
-        }
-        style={{ width: `${pct(context)}%` }}
-      />
-    </span>
-  );
-}
-
-function ContextBar({ context }: { context: number }) {
-  const threshold = (AUTOCOMPACT_TOKENS / MODEL_CONTEXT_WINDOW) * 100;
-  return (
-    <div className="relative h-2 w-full rounded-full bg-zinc-100">
-      <div
-        className={
-          "absolute inset-y-0 left-0 rounded-full transition-all " +
-          (context >= AUTOCOMPACT_TOKENS ? "bg-amber-400" : "bg-zinc-800")
-        }
-        style={{ width: `${pct(context)}%` }}
-      />
-      <div
-        className="absolute inset-y-0 w-px bg-amber-400"
-        style={{ left: `${threshold}%` }}
-        title={`Backend triggers summarization around ${formatTokenCount(AUTOCOMPACT_TOKENS)} tokens`}
-      />
-    </div>
-  );
-}
-
-const BREAKDOWN_COLORS = {
-  system: "bg-zinc-400",
-  user: "bg-sky-400",
-  assistant: "bg-violet-400",
-  tools: "bg-emerald-400",
-};
-
-function BreakdownSection({ breakdown }: { breakdown: ContextBreakdown }) {
-  const rows = [
-    {
-      label: "system + tools + skills",
-      tokens: BASE_CONTEXT_ESTIMATE,
-      color: BREAKDOWN_COLORS.system,
-      note: "fixed est.",
-    },
-    {
-      label: "your messages",
-      tokens: breakdown.userTokens,
-      color: BREAKDOWN_COLORS.user,
-    },
-    {
-      label: "assistant replies",
-      tokens: breakdown.assistantTokens,
-      color: BREAKDOWN_COLORS.assistant,
-    },
-    {
-      label: "tool calls + results",
-      tokens: breakdown.toolTokens,
-      color: BREAKDOWN_COLORS.tools,
-    },
-  ];
-  const total = breakdownTotal(breakdown);
-
-  return (
-    <div className="pb-2.5">
-      <div className="flex h-1.5 w-full gap-px overflow-hidden rounded-full">
-        {rows.map((row) => (
-          <div
-            key={row.label}
-            className={row.color}
-            style={{ width: `${(row.tokens / total) * 100}%` }}
-          />
-        ))}
-      </div>
-      <div className="flex flex-col gap-0.5 pt-1.5">
-        {rows.map((row) => (
-          <div
-            key={row.label}
-            className="flex items-baseline gap-1.5 font-mono text-xs"
-          >
-            <span className={`size-1.5 shrink-0 rounded-full ${row.color}`} />
-            <span className="text-zinc-600">{row.label}</span>
-            {row.note && (
-              <span className="text-[10px] text-zinc-400">{row.note}</span>
-            )}
-            <span className="ml-auto text-zinc-500">
-              ~{formatTokenCount(row.tokens)}
-            </span>
-          </div>
-        ))}
-      </div>
-    </div>
-  );
-}
-
-function TurnRow({ index, turn }: { index: number; turn: TokenTurn }) {
-  return (
-    <div className="flex items-baseline gap-2 font-mono text-xs">
-      <span className="w-6 shrink-0 text-zinc-500">#{index + 1}</span>
-      {turn.compacted && (
-        <span
-          title="Transcript summarized this turn"
-          className="text-amber-500"
-        >
-          ⟲
-        </span>
-      )}
-      <span className="text-zinc-800">
-        in {formatTokenCount(turnInputTokens(turn))}
-      </span>
-      <span className="text-zinc-500">
-        out {formatTokenCount(turn.completionTokens)}
-      </span>
-      <span className="ml-auto text-zinc-400">
-        w {formatTokenCount(turn.cacheCreationTokens)}
-      </span>
-    </div>
   );
 }
